+++ src/components/NavBar.tsx
@@ -1,7 +1,6 @@
 import Search from "./Search";
 import React, { useEffect, useState } from "react";
 import Link from "next/link";
-import Login from "./Login";
 import DropdownPane from "./DropdownPane";
 
 import { CgMenu } from "react-icons/cg";
@@ -11,28 +10,36 @@ import NavMenu from "./NavMenu";
 import { useRouter } from "next/router";
 import SortMenu from "./SortMenu";
 
-import { useSession } from "next-auth/react";
-import { useScroll } from "../hooks/useScroll";
-
 import { usePlausible } from "next-plausible";
 import { useMainContext } from "../MainContext";
 import FilterMenu from "./FilterMenu";
 import LoginProfile from "./LoginProfile";
 import useRefresh from "../hooks/useRefresh";
+import useNavBarScrollHelper from "../hooks/useNavBarScrollHelper";
+import { useWindowWidth } from "@react-hook/window-size";
+import { AiOutlineSearch } from "react-icons/ai";
 
 const NavBar = ({ toggleSideNav = 0 }) => {
   const context: any = useMainContext();
   const { invalidateKey, refreshCurrent, fetchingCount } = useRefresh();
   const plausible = usePlausible();
   const router = useRouter();
-
+  const windowWidth = useWindowWidth();
+  const [showSearch, setShowSearch] = useState(false);
   const [hidden, setHidden] = useState(false);
   const [allowHide, setallowHide] = useState(true);
   const [allowShow, setAllowShow] = useState(true);
   const [sidebarVisible, setSidebarVisible] = useState(false);
   //add some delay before navbar can be hidden again.. resolves some issues with immediate hide after navigation
   const [timeSinceNav, setTimeSinceNav] = useState(() => new Date().getTime());
-  const { scrollY, scrollDirection } = useScroll();
+
+  useNavBarScrollHelper({
+    allowHide,
+    allowShow,
+    setHidden,
+    timeSinceNav,
+    autoHideNav: context.autoHideNav,
+  });
 
   useEffect(() => {
     toggleSideNav && setSidebarVisible(true);
@@ -40,22 +47,6 @@ const NavBar = ({ toggleSideNav = 0 }) => {
       setSidebarVisible(false);
     };
   }, [toggleSideNav]);
-  useEffect(() => {
-      const now = new Date().getTime();
-      if (allowHide && now > timeSinceNav + 1000) {
-        if (scrollDirection === "down" || (!scrollY && allowShow)) {
-          setHidden(false);
-        } else if (scrollY && scrollY > 300 && scrollDirection === "up") {
-          setHidden(true);
-        } else if ((!scrollY || scrollY <= 300) && allowShow) {
-          setHidden(false);
-        }
-      } else {
-        if (allowShow) {
-          setHidden(false);
-        }
-      }
-  }, [scrollDirection, allowHide, scrollY]);
 
   useEffect(() => {
     setTimeSinceNav(() => new Date().getTime());
@@ -84,7 +75,7 @@ const NavBar = ({ toggleSideNav = 0 }) => {
       //setallowHide(true);
     };
   }, [router]);
-  
+
   useEffect(() => {
     if (context.mediaMode) {
       setTimeSinceNav(() => new Date().getTime());
@@ -113,7 +104,7 @@ const NavBar = ({ toggleSideNav = 0 }) => {
     return () => {
       window.removeEventListener("mousemove", updateMousePosition);
     };
-  }, [allowShow]);  
+  }, [allowShow]);
 
   const homeClick = () => {
     router?.route === "/" && invalidateKey(["feed", "HOME"], false); // setForceRefresh((p) => p + 1);
@@ -131,7 +122,7 @@ const NavBar = ({ toggleSideNav = 0 }) => {
         <SideNav visible={sidebarVisible} toggle={setSidebarVisible} />
         <nav className="flex flex-row items-center flex-grow h-full shadow-lg bg-th-background2 md:justify-between ">
           <CgMenu
-            className="w-10 h-10 cursor-pointer md:hidden"
+            className="flex-none w-10 h-10 cursor-pointer md:hidden"
             onClick={() => {
               setSidebarVisible((vis) => !vis);
               // plausible("sidenav");
@@ -159,7 +150,34 @@ const NavBar = ({ toggleSideNav = 0 }) => {
           <div className="hidden w-full h-full py-1.5 max-w-5xl md:block">
             <Search id={"subreddit search main"} />
           </div>
+           <div
+              className={
+                "flex-none  h-10 transition  duration-200 ease-in-out origin-top md:origin-top-right lg:origin-right " +
+                (showSearch
+                  ? " absolute top-[3.2rem] w-[90vw]  left-[5vw] md:left-[25vw] md:w-[50vw] lg:relative lg:top-auto lg:left-0  lg:w-[24rem] scale-x-100 "
+                  : " w-0 absolute lg:scale-x-0 scale-x-0 scale-y-0 lg:scale-y-100 opacity-0 ")
+              }
+            >
+              {showSearch && (
+                <Search
+                  id={"subreddit search main"}
+                  setShowSearch={windowWidth < 1024 ? setShowSearch : (a) => {}}
+                />
+              )}
+            </div>
           <div className="flex flex-row items-center justify-end h-full py-1.5 ml-auto mr-2 space-x-1 md:ml-2">
+          <button
+              disabled={windowWidth > 768}
+              aria-label="show search"
+              className={"flex md:hidden items-center justify-center flex-none w-10 h-full border border-transparent rounded-md outline-none hover:border-th-border "}
+              onClick={(e) => {
+                e.preventDefault();
+                e.stopPropagation();
+                setShowSearch((s) => !s);
+              }}
+            >
+              <AiOutlineSearch className="flex-none w-6 h-6" />
+            </button>
             <div className="w-20 h-full">
               <SortMenu hide={hidden} />
             </div>