+++ src/pages/settings.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import Settings from "../components/settings/Settings";
 
 const SettingsPage = () => {